--- app/controllers/concerns/microsoft_concern.rb
@@ -1,25 +1,0 @@
-module MicrosoftConcern
-  extend ActiveSupport::Concern
-
-  def microsoft_client
-    app_id = GlobalConfigService.load('AZURE_APP_ID', nil)
-    app_secret = GlobalConfigService.load('AZURE_APP_SECRET', nil)
-
-    ::OAuth2::Client.new(app_id, app_secret,
-                         {
-                           site: 'https://login.microsoftonline.com',
-                           authorize_url: 'https://login.microsoftonline.com/common/oauth2/v2.0/authorize',
-                           token_url: 'https://login.microsoftonline.com/common/oauth2/v2.0/token'
-                         })
-  end
-
-  private
-
-  def parsed_body
-    @parsed_body ||= Rack::Utils.parse_nested_query(@response.raw_response.body)
-  end
-
-  def base_url
-    ENV.fetch('FRONTEND_URL', 'http://localhost:3000')
-  end
-end
